Overhaul bot
This commit is a big overhaul of several changes.

The first change is the creation of a webhook to impersonate users and
archive messages in the archive channel. This is similar to what the
Sticker bot does to send messages as other users. There are some
heuristics to make attachments and embeds work properly.

The other change is to stop using the audit log to detect pinned
messages. Instead, just look at the most recent pinned message in the
pins, which are always ordered by most recent pin.

Finally, it makes the bot more robust by not pinning already-pinned
messages and checking making some permission checks more stringent.