Overhaul message archiving to use embeds
This whole thing overhauls the archiving function to use an embed, with
a bunch of heuristics for making it behave the same way as normal
messages w.r.t. automatically embedded images. Most notably, it tries to
set either the embed image or thumbnail as the first link in the
message.